refactor(architecture): move namespace-move batch-rename detection into compare/ (ADR-061)
Two of architecture/debt.yaml's six currently-failing debt-no-growth
baselines were fixable as a genuine responsibility move rather than a
line-count trim:

- diff_symbols_renames.py had drifted 97 lines past its adoption baseline.
  find_namespace_move_groups/emit_namespace_move_batches and their private
  helpers (_scope_components, _qualified_key_scope_components,
  _declaring_entity) were the whole namespace-move-substitution subsystem,
  used nowhere else in the file -- moved to the new
  abicheck/compare/namespace_move.py, matching the file's own ADR-061
  `compare` classification and the sibling compare/rename_ambiguity.py
  split. diff_symbols_renames.py now re-exports the two entry points so
  diff_symbols.py's existing public-surface re-export block is untouched.

- The synthetic ctor/dtor snapshot-key vocabulary (SYNTHETIC_CTOR_KEY_PREFIX,
  is_synthetic_ctor_key, is_synthetic_dtor_key) that namespace_move.py needs
  was previously reachable only through dumper_castxml.py (`extract`), which
  `compare` may not import per ADR-061 D1. Relocated to
  abicheck/model/synthetic_key.py -- the same "shared vocabulary both extract
  and compare need" pattern model/mangled_name.py already established --
  with extract/headers/castxml/names.py and dumper_castxml.py re-exporting
  it so every existing import path (including finding_identity_ctor_dtor.py)
  keeps working.

- tests/test_batch_rename_namespace_move.py had drifted 328 lines past its
  baseline; debt.yaml's own recorded target for it was "tests/unit/compare
  alongside namespace/rename comparison migration". Split by contract axis
  into tests/unit/compare/test_namespace_move.py (namespace-move grouping/
  emission + the small prefix-rename/destructor-noise regression it shared
  the file with) and tests/unit/compare/test_scope_component_parsing.py
  (qualified_name_scope_components/strip_trailing_top_level_parameter_list
  edge cases, which exercise diff_cxx_rules directly). Both new files land
  under the 1200-line test hard cap, so the debt.yaml entry is retired
  rather than retargeted.

architecture/debt.yaml: removed the retired test entry, lowered
diff_symbols_renames.py's baseline to its new (807-line) size.

No behavior change: mypy/ruff/format clean, and every test touching this
code (tests/unit/compare/, the cross-position-ambiguity/rename-ambiguity
suites, and the broader rename/namespace_move/diff_symbols/synthetic
keyword slice -- 736 tests total) passes unchanged.
