+++ abicheck/extract/headers/castxml/names.py
@@ -28,6 +28,13 @@ from __future__ import annotations
 import re
 from typing import Any
 
+from ....model.synthetic_key import (
+    _SYNTHETIC_DTOR_KEY_PREFIX as _SYNTHETIC_DTOR_KEY_PREFIX,
+    SYNTHETIC_CTOR_KEY_PREFIX as SYNTHETIC_CTOR_KEY_PREFIX,
+    is_synthetic_ctor_key as is_synthetic_ctor_key,
+    is_synthetic_dtor_key as is_synthetic_dtor_key,
+)
+
 
 def _parse_vtable_index(vi_str: str | None) -> int | None:
     """Parse vtable_index attribute, returning None for missing/invalid values."""
@@ -101,44 +108,6 @@ def _mangled_name_is_local_linkage(mangled: str) -> bool:
         if i < n and mangled[i] == "E":
             return False
     return False
-
-
-#: Prefix marking a snapshot key synthesized for a constructor overload whose
-#: real mangled name castxml omitted (see ``_CastxmlParser._function_mangled_name``).
-#: It is intentionally not a real ABI symbol, only a stable per-overload
-#: identity — ``diff_symbols._public_functions()`` reads this to exempt such
-#: entries from its ELF-export-set narrowing, which they could never pass (the
-#: key has no real exported symbol to match).
-SYNTHETIC_CTOR_KEY_PREFIX = "__abicheck_ctor__"
-
-
-def is_synthetic_ctor_key(key: str) -> bool:
-    """Whether *key* is a castxml constructor-overload synthetic identity."""
-    return key.startswith(SYNTHETIC_CTOR_KEY_PREFIX)
-
-
-#: Marker for a snapshot key synthesized for a destructor whose real mangled
-#: name castxml omitted (see ``_CastxmlParser._function_display_name`` and
-#: ``_function_mangled_name``'s ``return name`` fallback). A class has at
-#: most one destructor, so — unlike constructors — no per-overload prefix is
-#: needed: the synthesized "~ClassName" display name is itself already a
-#: stable, unique identity. It is intentionally not a real ABI symbol (a real
-#: Itanium destructor mangling always starts with ``_Z``, never ``~``), only
-#: a stable key — ``diff_symbols._public_functions()`` reads this the same
-#: way it already does :data:`SYNTHETIC_CTOR_KEY_PREFIX`/
-#: :func:`is_synthetic_ctor_key`, to exempt such entries from its
-#: ELF-export-set narrowing, which they could never pass. Without this, a
-#: real virtual destructor's PUBLIC visibility (``_ctor_or_dtor_visibility``)
-#: was necessary but not sufficient: it would still be silently dropped
-#: before reaching the diff whenever ELF metadata is present (Codex review,
-#: PR #582 — found after the destructor-visibility fix, via the same Phase 2
-#: parity gate).
-_SYNTHETIC_DTOR_KEY_PREFIX = "~"
-
-
-def is_synthetic_dtor_key(key: str) -> bool:
-    """Whether *key* is a castxml destructor synthetic identity."""
-    return key.startswith(_SYNTHETIC_DTOR_KEY_PREFIX)
 
 
 def _virtual_method_mangled_name(method_el: Any) -> str:
